T2-96

+++ src/main/java/com/crm/main/service/OrganizationRoleService.java
@@ -1,8 +1,10 @@
 package com.crm.main.service;
 
+import com.crm.main.enums.RoleType;
 import com.crm.main.persistance.entity.Organization;
 import com.crm.main.persistance.entity.OrganizationRole;
 import com.crm.main.persistance.repository.OrganizationRoleRepository;
+import com.crm.sharedlib.exception.ForbiddenException;
 import com.crm.sharedlib.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,12 +23,29 @@ public class OrganizationRoleService {
             Organization organization,
             Long roleId
     ) {
-        OrganizationRole role = new OrganizationRole();
+        return createOrganizationRole(
+                organization, roleId, RoleType.USER
+        );
+    }
 
-        role.setOrganization(organization);
-        role.setRoleId(roleId);
+    @Transactional
+    public OrganizationRole createAdminRoleForOrganization(
+            Organization organization,
+            Long roleId
+    ) {
+        return createOrganizationRole(
+                organization, roleId, RoleType.ADMIN
+        );
+    }
 
-        return repository.save(role);
+    @Transactional
+    public OrganizationRole createMemberRoleForOrganization(
+            Organization organization,
+            Long roleId
+    ) {
+        return createOrganizationRole(
+                organization, roleId, RoleType.MEMBER
+        );
     }
 
     public OrganizationRole getOrganizationRole(Organization organization, Long roleId) {
@@ -40,7 +59,26 @@ public class OrganizationRoleService {
 
     @Transactional
     public void deleteOrganizationRole(OrganizationRole organizationRole) {
+        RoleType roleType = organizationRole.getRoleType();
+
+        if (roleType == RoleType.ADMIN || roleType == RoleType.MEMBER) {
+            throw new ForbiddenException("You can't delete this role");
+        }
+
         repository.delete(organizationRole);
+    }
+
+    private OrganizationRole createOrganizationRole(
+            Organization organization, Long roleId,
+            RoleType roleType
+    ) {
+        OrganizationRole role = new OrganizationRole();
+
+        role.setOrganization(organization);
+        role.setRoleId(roleId);
+        role.setRoleType(roleType);
+
+        return repository.save(role);
     }
 
 }
